# lib/api-zod/src/generated/types/cardAdAccountsListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AdWallet Africa API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CardAdAccount } from './cardAdAccount';

export interface CardAdAccountsListResponse {
  adAccounts: CardAdAccount[];
}