# lib/api-zod/src/generated/types/koraCardDetails.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AdWallet Africa API specification
 * OpenAPI spec version: 0.1.0
 */

export interface KoraCardDetails {
  id: number;
  cardNumberMasked: string;
  last4: string;
  cardholderName: string;
  balance: number;
  spendingLimit: number;
  status: string;
  purpose: string;
  linkedAdAccounts: number;
  createdAt: Date;
}