# lib/api-zod/src/generated/types/koraCardDetailsCreated.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AdWallet Africa API specification
 * OpenAPI spec version: 0.1.0
 */

export interface KoraCardDetailsCreated {
  id: number;
  koraCardId: string;
  cardNumberMasked: string;
  expiry: string;
  cvv: string;
  balance: number;
  spendingLimit: number;
  status: string;
  purpose: string;
  createdAt: Date;
}