# lib/api-zod/src/generated/types/linkCardAdAccountInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AdWallet Africa API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LinkCardAdAccountInputAdPlatform } from './linkCardAdAccountInputAdPlatform';

export interface LinkCardAdAccountInput {
  adPlatform: LinkCardAdAccountInputAdPlatform;
  adAccountId: string;
  adAccountName: string;
}